refactoring: improve the name to be closer to http.Error
