+++ pkg/util/http.go
@@ -30,6 +30,7 @@ import (
 	"gopkg.in/yaml.v3"
 
 	"github.com/grafana/pyroscope/pkg/tenant"
+	httputil "github.com/grafana/pyroscope/pkg/util/http"
 )
 
 var defaultTransport http.RoundTripper = &http2.Transport{
@@ -73,7 +74,7 @@ func WriteYAMLResponse(w http.ResponseWriter, v interface{}) {
 
 	data, err := yaml.Marshal(v)
 	if err != nil {
-		WriteError(err, w)
+		httputil.Error(w, err)
 		return
 	}
 
@@ -355,7 +356,7 @@ func RenderHTTPResponse(w http.ResponseWriter, v interface{}, t *template.Templa
 	w.Header().Set("Content-Type", "text/html; charset=utf-8")
 	err := t.Execute(w, v)
 	if err != nil {
-		WriteError(err, w)
+		httputil.Error(w, err)
 	}
 }
 
@@ -365,7 +366,7 @@ func WriteJSONResponse(w http.ResponseWriter, v interface{}) {
 
 	data, err := json.Marshal(v)
 	if err != nil {
-		WriteError(err, w)
+		httputil.Error(w, err)
 		return
 	}
 
@@ -386,7 +387,7 @@ func AuthenticateUser(on bool) middleware.Interface {
 			}
 			_, ctx, err := user.ExtractOrgIDFromHTTPRequest(r)
 			if err != nil {
-				WriteErrorWithStatus(err, w, http.StatusUnauthorized)
+				httputil.ErrorWithStatus(w, err, http.StatusUnauthorized)
 				return
 			}
 			next.ServeHTTP(w, r.WithContext(ctx))
